scripts/requirements-poster.txt
# For scripts/capture_hero_poster.py
playwright>=1.40
pillow>=9.0
